# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Synque Express Ticketing API",
    "version": "0.1.0",
    "description": "Complete ticketing and event management API with user profile management, promotions, and cancellation support"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "https://api.ticketing.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/orders": {
      "post": {
        "tags": [
          "Commerce"
        ],
        "summary": "Create new order",
        "description": "Create a new order for one or more products",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "user_id",
                  "items",
                  "channel_id",
                  "out_trade_no"
                ],
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "description": "ID of the user placing the order"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "product_id": {
                          "type": "integer"
                        },
                        "qty": {
                          "type": "integer",
                          "minimum": 1
                        }
                      }
                    }
                  },
                  "channel_id": {
                    "type": "integer",
                    "description": "Sales channel identifier"
                  },
                  "out_trade_no": {
                    "type": "string",
                    "description": "External transaction number for idempotency"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order_id": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "paid",
                        "cancelled"
                      ]
                    },
                    "amounts": {
                      "type": "object",
                      "properties": {
                        "subtotal": {
                          "type": "number",
                          "format": "decimal"
                        },
                        "discount": {
                          "type": "number",
                          "format": "decimal"
                        },
                        "total": {
                          "type": "number",
                          "format": "decimal"
                        }
                      }
                    },
                    "out_trade_no": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Order with this out_trade_no already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ota/orders": {
      "get": {
        "tags": [
          "OTA Integration"
        ],
        "summary": "List all confirmed OTA orders",
        "description": "Retrieve paginated list of orders created through OTA platform",
        "security": [
          {
            "otaApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            },
            "description": "Maximum number of orders to return"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of orders to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "order_id": {
                            "type": "string",
                            "example": "ORD-1762273241549"
                          },
                          "customer_name": {
                            "type": "string",
                            "example": "Maria Garcia"
                          },
                          "customer_email": {
                            "type": "string",
                            "example": "maria.garcia@example.com"
                          },
                          "total_amount": {
                            "type": "string",
                            "example": "382.00"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "confirmed",
                              "completed",
                              "cancelled"
                            ]
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "confirmation_code": {
                            "type": "string",
                            "example": "CONF-1762273241549"
                          }
                        }
                      }
                    },
                    "total_count": {
                      "type": "integer",
                      "example": 2
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/my/orders": {
      "get": {
        "tags": [
          "Commerce"
        ],
        "summary": "List user's orders",
        "description": "Get a list of all orders for the authenticated user (alternative endpoint)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    },
                    "total_count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "integer",
            "description": "Unique order identifier"
          },
          "user_id": {
            "type": "integer",
            "description": "User who placed the order"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PENDING_PAYMENT",
              "PAID",
              "CANCELLED",
              "REFUNDED",
              "PARTIALLY_REFUNDED"
            ],
            "description": "Current order status"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "product_id": {
                  "type": "integer"
                },
                "qty": {
                  "type": "integer"
                }
              }
            },
            "description": "Order items"
          },
          "channel_id": {
            "type": "integer",
            "description": "Sales channel identifier"
          },
          "out_trade_no": {
            "type": "string",
            "description": "External transaction number"
          },
          "amounts": {
            "type": "object",
            "properties": {
              "subtotal": {
                "type": "number",
                "format": "decimal"
              },
              "discount": {
                "type": "number",
                "format": "decimal"
              },
              "total": {
                "type": "number",
                "format": "decimal"
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Order creation timestamp"
          },
          "paid_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Payment completion timestamp"
          },
          "refund_amount": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "refund_status": {
            "type": "string",
            "enum": [
              "none",
              "partial",
              "full"
            ],
            "nullable": true
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authenticated endpoints"
      },
      "otaApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for OTA platform integration"
      }
    }
  },
...
}